requirements.txt
# Python packages - install with: pip install -r requirements.txt

Flask==3.0.0
mysql-connector-python==8.2.0

# Frontend tools used (loaded via CDN, no install needed):
# - Bootstrap 5.3  → https://cdn.jsdelivr.net/npm/bootstrap@5.3.0
# - jQuery 3.6     → https://code.jquery.com/jquery-3.6.0.min.js
